Reorganize tests, add code coverage
