+++ .github/dependabot.yml
@@ -10,6 +10,9 @@ updates:
     directory: "/"
     schedule:
       interval: "weekly"
+    ignore:
+      - dependency-name: "*"
+        update-types: ["version-update:semver-patch", "version-update:semver-minor"]
 
   # Maintain dependencies for npm
   - package-ecosystem: "npm"
