+++ .github/workflows/phpunit.yml
@@ -18,6 +18,7 @@ jobs:
       with:
         php-version: 8.3
         extensions: dom, curl, libxml, mbstring, zip, fileinfo
+        coverage: xdebug
     - name: Copy .env
       run: php -r "file_exists('.env') || copy('.env.example', '.env');"
     - name: Install Dependencies
@@ -36,4 +37,10 @@ jobs:
       env:
         DB_CONNECTION: sqlite
         DB_DATABASE: database/database.sqlite
-      run: php artisan test
+      run: php artisan test --coverage
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v4
+      with:
+        verbose: true
+      env:
+        CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
