+++ composer.json
@@ -47,8 +47,10 @@
         "laravel/sail": "^1.12",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^8.1",
+        "pestphp/pest": "^2.34",
+        "pestphp/pest-plugin-drift": "^2.6",
+        "pestphp/pest-plugin-laravel": "^2.4",
         "phpstan/phpstan": "^1.10",
-        "phpunit/phpunit": "^11.0",
         "spatie/laravel-ignition": "^2.0",
         "spatie/laravel-typescript-transformer": "^2.1"
     },
@@ -90,7 +92,8 @@
         "preferred-install": "dist",
         "sort-packages": true,
         "allow-plugins": {
-            "php-http/discovery": true
+            "php-http/discovery": true,
+            "pestphp/pest-plugin": true
         }
     },
     "minimum-stability": "stable",
