+++ tests/Feature/Other/ExampleTest.php
@@ -1,0 +1,7 @@
+<?php
+
+it('returns a successful response', function () {
+    $response = $this->get('/lt');
+
+    $response->assertStatus(200);
+});
