+++ tests/Unit/ExampleTest.php
@@ -1,20 +1,7 @@
 <?php
 
-namespace Tests\Unit;
-
 use PHPUnit\Framework\Attributes\CoversNothing;
-use PHPUnit\Framework\TestCase;
 
-#[CoversNothing]
-class ExampleTest extends TestCase
-{
-    /**
-     * A basic test example.
-     *
-     * @return void
-     */
-    public function test_true_is_true()
-    {
-        $this->assertTrue(true);
-    }
-}
+test('true is true', function () {
+    expect(true)->toBeTrue();
+});
